Resign from storing EC indices on atoms
Previously, the EC indices were stored directly on atoms as an
additional property. This approach turned out to be a little cumbersome
as far as managing those indices for needs of Lynch-Willett algorithm
was concerned. Thus, I changed it and now EC indices are stored
'centrally' as a class attribute (a list).

Also, to make managing EC indices more flexible, functions which
calculates them do NOT update them implicitly, they just return EC
values to the caller. Updating is done by calling a dedicated function.