Fix bugs in calc_init_ecs()
Method calculating initial EC values had to bugs. Firstly, the default
value of 'index_type' was spelled incorrectly. Thus, calling the method
with no option specified was triggering the fallback mechanism instead
of selecting the appropriate method. Secondly, when invoked, the
fallback mechanism was returning method name instead of method itself.

I corrected the spelling error and fixed fallback mechanism to make it
return what it was supposed to.