Modify stopping condition slightly
A part of stopping condition checking if EC order equal the size of the
the largest reactant/product part. In special cases, when the part
was a single atom it (e.g. has size 1), it caused the script to enter an
infinite loop as EC order was at least 2 at this point. Thus, I modified
it to check whether the order is greater or equal than the aforementioned
size.