Add option 'verbose' to the Reaction class
Printing out match radius and atom mappings in each iteration of LW
algorithm is valuable for testing and demonstrating purposes. It is
complete waste of time though, if script is running in a batch mode
(e.g. analyzing a reaction database). Thus, I added a boolean option
'verbose' (defaults to False) to the Reaction class allowing to turn
on/off its verbose mode.