Fix core inclusion check
Core's method checking if a core contains other one was varying the
orderings of both reactant and retorns.  Firstly, it was inefficient as
it rendered linear operation of checking whether there is a match
between reactants and retros N^2. Secondly, due to not checking all
possible matches, it kept returning false negatives.  I corrected it by
removing line varying order of retrons.