Refactor method calculating similarity distance
Method calculating similarity between patterns was always returning (the
smallest) arithmetic average of their fragments' distances. This
approach was resulting in returning values less than 1 even for cases in
which incompatible fragments (i.e. having no MCS) existed.

Such a behavior was inconsistent with the function checking
the relation of inclusion between two patterns as it considers pattern
A to contain pattern B only if there exists an ordering of A's fragments
which contains (in terms of subgraph isomorphism) all B's fragments
simultaneously.

Thus, I modified the method to return an arithmetic average of partial
distances only if all of them have non empty MCSes. Otherwise, if there
is no such ordering, it returns 1.